# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mini-Lago API (Numaris-native)",
    "version": "2.0.0",
    "description": "Motor de facturación domain-specific para Numaris. Entidades: Customer, Tax, Service (colapsa Plan+Sub+Charge+BM+AddOn), Unit (materializada), EventLog (audit), Invoice, CreditNote.\n\n**Auth**: Bearer en todos los `/api/v1/*` excepto health/openapi/discovery. Callbacks NetSuite usan HMAC-SHA256 (`X-NetSuite-Signature`).\n\n**Idempotency** en POST /invoices y POST /credit_notes vía header `Idempotency-Key` + `metadata.idempotency_key`."
  },
  "servers": [
    { "url": "https://billing2-production.up.railway.app", "description": "Railway production" },
    { "url": "http://localhost:3000", "description": "Local dev" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer" },
      "NetSuiteSignature": { "type": "apiKey", "in": "header", "name": "X-NetSuite-Signature" }
    },
    "schemas": {
...
      "Tax": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "code": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "rate": { "type": "number", "example": 16.0 }
        }
      },
      "TaxWriteable": {
        "type": "object",
        "required": ["code", "name", "rate"],
        "properties": {
          "code": { "type": "string", "example": "iva-mx-16" },
          "name": { "type": "string", "example": "IVA México" },
          "description": { "type": "string" },
          "rate": { "oneOf": [{ "type": "string" }, { "type": "number" }], "example": "16" }
        }
      },
      "Service": {
        "type": "object",
        "description": "Numaris-native: 1 customer ↔ N services. Cada service tiene su propia tarifa mensual por unidad + setup one-off.",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "code": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "customer_id": { "type": "string", "format": "uuid" },
          "currency": { "type": "string" },
          "monthly_unit_amount_cents": { "type": "integer" },
          "setup_unit_amount_cents": { "type": "integer" },
          "status": { "type": "string", "enum": ["pending", "active", "terminated"] },
          "billing_time": { "type": "string", "enum": ["calendar", "anniversary"] },
          "subscription_at": { "type": "string", "format": "date-time" },
          "started_at": { "type": ["string", "null"], "format": "date-time" },
          "current_billing_period_started_at": { "type": ["string", "null"], "format": "date-time" },
          "current_billing_period_ending_at": { "type": ["string", "null"], "format": "date-time" },
          "taxes": { "type": "array", "items": { "$ref": "#/components/schemas/Tax" }, "description": "Si vacío, hereda taxes del customer." },
          "metadata": { "type": "object" }
        }
      },
      "ServiceWriteable": {
        "type": "object",
        "required": ["code", "name", "customer_external_id"],
        "properties": {
          "code": { "type": "string", "example": "combustible-foo" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "customer_external_id": { "type": "string" },
          "currency": { "type": "string", "description": "Default: customer.currency." },
          "monthly_unit_amount_cents": { "type": "integer", "minimum": 0, "example": 45000 },
          "setup_unit_amount_cents": { "type": "integer", "minimum": 0, "example": 120000 },
          "billing_time": { "type": "string", "enum": ["calendar", "anniversary"], "default": "calendar" },
          "subscription_at": { "type": "string", "format": "date-time", "description": "Si es futuro → status=pending." },
          "tax_codes": { "type": "array", "items": { "type": "string" }, "description": "Si vacío, usa los del customer." },
          "metadata": { "type": "object" }
        }
      },
...
      "UnitWriteable": {
        "type": "object",
        "required": ["service_code", "external_id"],
        "properties": {
          "service_code": { "type": "string" },
          "external_id": { "type": "string" },
          "label": { "type": "string" },
          "active_from": { "type": "string", "format": "date-time" },
          "metadata": { "type": "object" }
        }
      },
...
    }
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/api/v1/taxes": {
      "post": { "summary": "Create tax", "tags": ["taxes"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "tax": { "$ref": "#/components/schemas/TaxWriteable" } } } } } }, "responses": { "200": { "description": "Tax" } } },
      "get": { "summary": "List taxes", "tags": ["taxes"], "responses": { "200": { "description": "Lista" } } }
    },
    "/api/v1/taxes/{code}": {
      "get": { "summary": "Read tax", "tags": ["taxes"], "parameters": [{ "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Tax" } } }
    },
    "/api/v1/services": {
      "post": {
        "summary": "Create service (colapsa Plan+Sub+Charge+BM+AddOn)",
        "tags": ["services"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "service": { "$ref": "#/components/schemas/ServiceWriteable" } } } } } },
        "responses": { "200": { "description": "Service", "content": { "application/json": { "schema": { "type": "object", "properties": { "service": { "$ref": "#/components/schemas/Service" } } } } } } }
      },
      "get": {
        "summary": "List services",
        "tags": ["services"],
        "parameters": [
          { "name": "customer_external_id", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "active", "terminated"] } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } }
        ],
        "responses": { "200": { "description": "Lista" } }
      }
    },
    "/api/v1/services/{code}": {
      "get": { "summary": "Read service", "tags": ["services"], "parameters": [{ "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Service" } } },
      "delete": { "summary": "Delete service (409 si tiene invoices — usa terminate en su lugar)", "tags": ["services"], "parameters": [{ "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" }, "409": { "description": "service_has_invoices" } } }
    },
    "/api/v1/services/{code}/terminate": {
      "post": { "summary": "Terminate service + todas sus units activas", "tags": ["services"], "parameters": [{ "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Service terminated" } } }
    },
    "/api/v1/units": {
      "post": { "summary": "Create unit (alternativa a POST /events con operation_type=add)", "tags": ["units"], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "unit": { "$ref": "#/components/schemas/UnitWriteable" } } } } } }, "responses": { "200": { "description": "Unit" } } },
      "get": {
        "summary": "List units",
        "tags": ["units"],
        "parameters": [
          { "name": "service_code", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["active", "terminated"] } }
        ],
        "responses": { "200": { "description": "Lista" } }
      }
    },
    "/api/v1/units/{id}": {
      "get": { "summary": "Read unit", "tags": ["units"], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Unit" } } },
      "patch": { "summary": "Update unit (label, active_to, metadata)", "tags": ["units"], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Unit" } } }
    },
...
  }
}